convert: Allow capsule types to overload conversions to and from
In the previous commit we allowed capsule types to override some basic
operations that callers reasonably expect all cty types to be able to
perform reasonably.

This follow-up gives similar treatment to the conversion operations.
Although conversion is not a core feature of the cty package, the convert
package is commonly used by callers as part of coercing values to fit
expected types, and so it's helpful to be able to define conversions to
and from capsule types in order to avoid re-inventing the whole conversion
mechanism in a calling application.

This is an unusual CapsuleOps field in that cty itself doesn't use it at
all, but the convert package can pull it out and make use of it
nonetheless.